refactor: authのフォルダ見直し
